Added only-retaliate-skill plugins + updates
Added player and AI versions of only-retaliate that work with a custom skill with the keyword "onlyRetaliate". Added SkillControl-getPossessionCustomSkillNonWeapon.js to work alongside the aforementioned custom skill versions. These are meant to achieve the same functionality of disabling the attack command for certain units/weapons just like the custom parameter versions, without having to add a custom parameter in the first place. 

Also added extra comments to custom parameter version of only-retaliate to inform how disabling an AI unit from initiating an attack can be done in-engine without having to use the AI plugins for that purpose.